Add files via upload
Hiring process is the fundamental and the most important
function of a company.
MNCs get to know about the major underlying trends
about the hiring process. Trends such as- number of
rejections, number of interviews, types of jobs, vacancies
etc. are important for a company to analyse before hiring
freshers or any other individual.